refactor: errorState 이동
errorState는 dto가 아니라 network 관련 데이터이므로 해당 패키지로 이동
