+++ android/app/src/main/java/app/priceguard/data/network/ProductErrorState.kt
@@ -1,4 +1,4 @@
-package app.priceguard.data.dto
+package app.priceguard.data.network
 
 enum class ProductErrorState {
     PERMISSION_DENIED,
