+++ android/app/src/main/java/app/priceguard/data/network/ProductRepositoryResult.kt
@@ -1,7 +1,5 @@
 package app.priceguard.data.network
 
-import app.priceguard.data.dto.ProductErrorState
-
 sealed class ProductRepositoryResult<out T> {
 
     data class Success<out T>(val data: T) : ProductRepositoryResult<T>()
