+++ android/app/src/main/java/app/priceguard/ui/additem/link/RegisterItemLinkViewModel.kt
@@ -2,9 +2,9 @@ package app.priceguard.ui.additem.link
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import app.priceguard.data.dto.ProductErrorState
 import app.priceguard.data.dto.verify.ProductVerifyRequest
 import app.priceguard.data.dto.verify.ProductVerifyResult
+import app.priceguard.data.network.ProductErrorState
 import app.priceguard.data.network.ProductRepositoryResult
 import app.priceguard.data.repository.ProductRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
