+++ android/app/src/main/java/app/priceguard/ui/detail/ProductDetailViewModel.kt
@@ -3,9 +3,9 @@ package app.priceguard.ui.detail
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import app.priceguard.data.GraphDataConverter
-import app.priceguard.data.dto.ProductErrorState
 import app.priceguard.data.graph.ProductChartData
 import app.priceguard.data.graph.ProductChartDataset
+import app.priceguard.data.network.ProductErrorState
 import app.priceguard.data.network.ProductRepositoryResult
 import app.priceguard.data.repository.ProductRepository
 import app.priceguard.materialchart.data.GraphMode
